Further reorganize workspace-related code
