+++ internal/app/forklift/caching-models.go
@@ -1,0 +1,9 @@
+package forklift
+
+import (
+	"github.com/PlanktoScope/forklift/pkg/pallets"
+)
+
+type FSCache struct {
+	FS pallets.PathedFS
+}
