+++ internal/app/forklift/workspace-models.go
@@ -8,18 +8,7 @@ type FSWorkspace struct {
 	FS pallets.PathedFS
 }
 
-// Env
-
-const currentEnvDirName = "env"
-
-type FSEnv struct {
-	FS pallets.PathedFS
-}
-
-// Cache
-
-const cacheDirName = "cache"
-
-type FSCache struct {
-	FS pallets.PathedFS
-}
+const (
+	currentEnvDirName = "env"
+	cacheDirName      = "cache" // TODO: have a cache of repos, a cache of envs
+)
